# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Primeiro Prontuário API
  version: 1.0.0
  description: |
    Contrato completo do MVP acadêmico para um único médico.
    Use somente dados fictícios ou anonimizados.
  license:
    name: Não licenciado - uso acadêmico demonstrativo
    identifier: LicenseRef-Proprietary
servers:
  - url: https://prontuario.local
    description: Cluster Kubernetes local com TLS confiado apenas na máquina de demonstração
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/patients:
    get:
      tags: [Pacientes]
      operationId: searchPatients
      summary: Pesquisar pacientes
      parameters:
        - name: fullName
          in: query
          schema:
            type: string
        - name: motherName
          in: query
          schema:
            type: string
        - name: cpf
          in: query
          schema:
            type: string
        - name: phone
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PatientStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: sort
          in: query
          description: campo,direção; campos permitidos estão no enum
          schema:
            type: string
            default: fullName,asc
            enum:
              - id,asc
              - id,desc
              - fullName,asc
              - fullName,desc
              - motherName,asc
              - motherName,desc
              - birthDate,asc
              - birthDate,desc
              - cpf,asc
              - cpf,desc
              - phone,asc
              - phone,desc
              - email,asc
              - email,desc
              - status,asc
              - status,desc
      responses:
        "200":
          description: Página de pacientes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Pacientes]
      operationId: createPatient
      summary: Cadastrar paciente
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Paciente criado
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/patients/{patientId}:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [Pacientes]
      operationId: getPatient
      summary: Consultar paciente
      responses:
        "200":
          description: Paciente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    put:
      tags: [Pacientes]
      operationId: updatePatient
      summary: Substituir os dados mutáveis do paciente
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
      responses:
        "200":
          description: Paciente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/patients/{patientId}/status:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    patch:
      tags: [Pacientes]
      operationId: changePatientStatus
      summary: Ativar ou inativar paciente
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePatientStatusRequest"
      responses:
        "200":
          description: Paciente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessão server-side; HttpOnly, SameSite=Lax e Secure em HTTPS
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        Token devolvido por GET /api/v1/auth/csrf e associado ao cookie
        XSRF-TOKEN do mesmo cliente. Exigido em mutações autenticadas.
  headers:
    CorrelationId:
      description: UUID de correlação aceito ou gerado pela API
      schema:
        type: string
        format: uuid
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Problem Details seguro; status e tipo dependem da falha
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    PatientStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    CreatePatientRequest:
      type: object
      required: [fullName, motherName, birthDate, cpf, phone]
      additionalProperties: false
      properties:
        fullName:
          type: string
          minLength: 1
        motherName:
          type: string
          minLength: 1
        birthDate:
          type: string
          format: date
        cpf:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: [string, "null"]
        address:
          type: [string, "null"]
        emergencyContact:
          type: [string, "null"]
        insurance:
          type: [string, "null"]
        allergies:
          type: [string, "null"]
        notes:
          type: [string, "null"]
    UpdatePatientRequest:
      allOf:
        - $ref: "#/components/schemas/CreatePatientRequest"
        - type: object
          required: [version]
          properties:
            version:
              type: integer
              format: int64
              minimum: 0
    ChangePatientStatusRequest:
      type: object
      required: [status, version]
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/PatientStatus"
        version:
          type: integer
          format: int64
          minimum: 0
    Patient:
      type: object
      required:
        - id
        - fullName
        - motherName
        - birthDate
        - cpf
        - phone
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        motherName:
          type: string
        birthDate:
          type: string
          format: date
        cpf:
          type: string
        phone:
          type: string
        email:
          type: [string, "null"]
        address:
          type: [string, "null"]
        emergencyContact:
          type: [string, "null"]
        insurance:
          type: [string, "null"]
        allergies:
          type: [string, "null"]
        notes:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/PatientStatus"
        version:
          type: integer
          format: int64
    PatientSummary:
      type: object
      required: [id, fullName, motherName, birthDate, cpf, phone, status]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        motherName:
          type: string
        birthDate:
          type: string
          format: date
        cpf:
          type: string
        phone:
          type: string
        email:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/PatientStatus"
    PatientPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/PatientSummary"
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [type, title, status, detail, correlationId]
      properties:
        type:
          type: string
          format: uri
          examples: [urn:problem:invalid-request]
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        correlationId:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"